Exclude headnote and otherHeadnote
RISDEV-6917
